+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add ability to archive/unarchive runs (gorarakelyan)
 - Enable search by run attributes (gorarakelyan)
 - Add `is not` keyword to AimQL (gorarakelyan)